implementation of dark theme
