+++ src/reducers/index.js
@@ -10,6 +10,7 @@ import widReducer from './widpane';
 import appReducer from './apps';
 import menusReducer from './menu';
 import globalReducer from './globals';
+import settReducer from './settings';
 
 const allReducers = combineReducers({
   wallpaper: wallReducer,
@@ -21,6 +22,7 @@ const allReducers = combineReducers({
   apps: appReducer,
   menus: menusReducer,
   globals: globalReducer,
+  setting: settReducer
 });
 
 var store = createStore(
